+++ packages/go_router_builder/example/lib/json_example.g.dart
@@ -1,0 +1,67 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: always_specify_types, public_member_api_docs
+
+part of 'json_example.dart';
+
+// **************************************************************************
+// GoRouterGenerator
+// **************************************************************************
+
+List<RouteBase> get $appRoutes => [
+      $homeRoute,
+    ];
+
+RouteBase get $homeRoute => GoRouteData.$route(
+      path: '/',
+      name: 'Home',
+      factory: $HomeRouteExtension._fromState,
+      routes: [
+        GoRouteData.$route(
+          path: 'json',
+          factory: $JsonRouteExtension._fromState,
+        ),
+      ],
+    );
+
+extension $HomeRouteExtension on HomeRoute {
+  static HomeRoute _fromState(GoRouterState state) => const HomeRoute();
+
+  String get location => GoRouteData.$location(
+        '/',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+extension $JsonRouteExtension on JsonRoute {
+  static JsonRoute _fromState(GoRouterState state) => JsonRoute(
+        (String json0) {
+          return JsonExample.fromJson(
+              jsonDecode(json0) as Map<String, dynamic>);
+        }(state.uri.queryParameters['json']!),
+      );
+
+  String get location => GoRouteData.$location(
+        '/json',
+        queryParams: {
+          'json': jsonEncode(json.toJson()),
+        },
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}